% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_dir}
\alias{cache_dir}
\title{Cache Directory}
\usage{
cache_dir()
}
\value{
Directory path
}
\description{
Location of weathercan cache (or where the cache will be if created).
}
\examples{
cache_dir()
}
