% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{cache_remove}
\alias{cache_remove}
\title{Remove cached directory and contents}
\usage{
cache_remove()
}
\value{
nothing but removes directory
}
\description{
This is a helper function to remove weathercan's cache directory.
This folder contains updated stations inventory lists and the full dataset of
Climate normals for 1991-2010.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
cache_remove()
\dontshow{\}) # examplesIf}
}
