% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{codes}
\alias{codes}
\title{Meaning of climate normal 'codes'}
\format{A data frame with 4 rows and 2 variables:
\describe{
\item{code}{Code}
\item{meaning}{Explanation of the code}
}}
\usage{
codes
}
\description{
A reference dataset containing \code{codes} matched to their meaning. Data
downloaded using the \code{normals_dl()} function contains columns indicating
\code{code}. These are presented here for interpretation.
}
\keyword{datasets}
