% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{envirocan-package}
\alias{envirocan-package}
\alias{envirocan}
\title{Easy downloading of weather data from Environment Canada}
\description{
Extended documentation coming soon... see index for details.
}
\references{
Environment Canada \url{}.
}
