% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finches}
\alias{finches}
\title{RFID Data on finch visits to feeders}
\format{An object of class \code{data.frame} with 16595 rows and 6 columns.}
\usage{
finches
}
\description{
An example dataset of finch RFID data for interpolation
}
\keyword{datasets}

