% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flags}
\alias{flags}
\title{Meaning of coded 'flags'}
\format{
A data frame with 16 rows and 2 variables:
\describe{
\item{code}{Flag code}
\item{meaning}{Explanation of the code}
}
}
\usage{
flags
}
\description{
A reference dataset containing 'flags' matched to their meaning. Data
downloaded using the \code{weather_dl()} function contains columns indicating
'flags' these codes are presented here for interpretation.
}
\keyword{datasets}
