% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glossary}
\alias{glossary}
\title{Glossary of units and terms}
\format{
A data frame with 77 rows and 5 variables:
\describe{
\item{interval}{Data interval type, 'hour', 'day', or 'month'.}
\item{ECCC_name}{Original column name when downloaded directly from ECCC}
\item{weathercan_name}{R-compatible name given when downloaded with the
\code{weather_dl()} function using the default argument \code{format =
  TRUE}.}
\item{units}{Units of the measurement.}
\item{ECCC_ref}{Link to the glossary or reference page on the ECCC
website.}
}
}
\usage{
glossary
}
\description{
A reference dataset matching information on columns in data downloaded using
the \code{weather_dl()} function. Indicates the units of the data, and
contains a link to the ECCC glossary page explaining the measurement.
}
\keyword{datasets}
