% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{glossary_normals}
\alias{glossary_normals}
\title{Glossary of terms for Climate Normals}
\format{A data frame with 18 rows and 3 variables:
\describe{
\item{ECCC_name}{Original measurement type from ECCC}
\item{weathercan_name}{R-compatible name given when downloaded with the
\code{normals_dl()} function}
\item{description}{Description of the measurement type from ECCC}
}}
\usage{
glossary_normals
}
\description{
A reference dataset matching information on columns in climate normals data
downloaded using the \code{normals_dl()} function. Indicates the names and
descriptions of different data measurements.
}
\keyword{datasets}
