% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kamloops}
\alias{kamloops}
\title{Hourly weather data for Kamloops downloaded with \code{weather()}}
\format{An example dataset of hourly weather data for Kamloops:
\describe{
  \item{station_name}{Station name}
  \item{station_id}{Environment Canada's station ID number. Required for downloading station data.}
  \item{prov}{Province}
  \item{lat}{Latitude of station location}
  \item{lon}{Longitude of station location}
  \item{date}{Date}
  \item{time}{Time}
  \item{year}{Year}
  \item{month}{Month}
  \item{day}{Day}
  \item{hour}{Hour}
  \item{qual}{Data quality}
  \item{weather}{Weather}
  \item{hmdx}{Humidex}
  \item{hmdx_flag}{Humidex data flag}
  \item{pressure}{Pressure}
  \item{pressure_flag}{Pressure data flag}
  \item{rel_hum}{Relative humidity}
  \item{rel_hum_flag}{Relative humidity data flag}
  \item{temp}{Temperature}
  \item{temp_dew}{Temperature of Dew Point}
  \item{temp_dew_flag}{Temperature of Dew Point data flag}
  \item{visib}{Visibility}
  \item{visib_flag}{Visibility data flag}
  \item{wind_chill}{Wind Chill}
  \item{wind_chill_flag}{Wind Chill flag}
  \item{wind_dir}{Wind Direction}
  \item{wind_dir_flag}{wind Direction Flag}
  \item{elev}{Elevation}
  \item{climat_id}{Climate identifier}
  \item{WMO_id}{World Meteorological Organization Identifier}
  \item{TC_id}{Transport Canada Identifier}
  \item{climate_id}{Climate ID number}
  \item{WMO_id}{Climate ID number}
  \item{TC_id}{Climate ID number}
  \item{elev}{Elevation of station location}
  \item{interval}{Interval of the data measurements ('hour', 'day', 'month')}
  \item{start}{Starting year of station data}
  \item{end}{Ending year of station data}
}}
\source{
\url{http://climate.weather.gc.ca/index_e.html}
}
\usage{
kamloops
}
\description{
Hourly weather data for Kamloops downloaded with \code{weather()}
}
\keyword{datasets}
