% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kamloops_day}
\alias{kamloops_day}
\title{Daily weather data for Kamloops downloaded with \code{weather()}}
\format{An object of class \code{data.frame} with 182 rows and 36 columns.}
\usage{
kamloops_day
}
\description{
An example dataset of daily weather data for Kamloops
}
\keyword{datasets}

