% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normals_cached.R
\name{normals_file}
\alias{normals_file}
\title{Location of the cached normals data}
\usage{
normals_file(normals_years = "1991-2020", type = "normals")
}
\arguments{
\item{normals_years}{Character. Years to load. Currently only \code{1991-2020} is available.}

\item{type}{Character. Data type to load, one of "normals", or "meta" (the
composite station inventory).}
}
\value{
Character file path.
}
\description{
Returns the \emph{expected} file path of the location of cached normals data. Only the most
current normals data are provided as a full data download by ECCC, so only
these normals are cached locally. Note that if you haven't downloaded
the normals files yet (call \code{normals_dl(...)}) these files will not exist.
}
\examples{
normals_file()
}
