% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{normals_measurement_types}
\alias{normals_measurement_types}
\title{List of climate normals measurements and types for each set of normals}
\format{
A data frame with 158 rows and 3 variables:
\describe{
\item{normals}{Year range of climate normals}
\item{measurement_type}{Type of measurement (relevant only for 1990-2020)}
\item{measurement}{Climate normals measurement available for this station}
}
}
\usage{
normals_measurement_types
}
\description{
A data frame listing the climate normals measurements classified by
measurement_type available for each set of climate normals. This is very
similar to \code{normals_measurements} and just omits the stations.
}
\keyword{datasets}
