% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{normals_measurements}
\alias{normals_measurements}
\title{List of climate normals measurements for each station}
\format{
A data frame with 113,325 rows and 5 variables:
\describe{
\item{prov}{Province}
\item{station_name}{Station Name}
\item{climate_id}{Climate ID}
\item{normals}{Year range of climate normals}
\item{measurement}{Climate normals measurement available for this station}
}
}
\usage{
normals_measurements
}
\description{
A data frame listing the climate normals measurements available for each
station.
}
\keyword{datasets}
