% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pg}
\alias{pg}
\title{Hourly weather data for Prince George downloaded with \code{weather()}}
\format{An object of class \code{data.frame} with 4368 rows and 35 columns.}
\usage{
pg
}
\description{
An example dataset of hourly weather data for Prince George
}
\keyword{datasets}

