% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations_all}
\alias{stations_all}
\title{Get available stations}
\usage{

  stations_all(url = "ftp://client_climate@ftp.tor.ec.gc.ca/Pub/Get_More_Data_Plus_de_donnees/Station Inventory EN.csv",
  skip = NULL)
}
\arguments{
\item{url}{Character. Url from which to grab the station information}

\item{skip}{Numeric. Number of lines to skip at the beginning of the csv. If
NULL, automatically derived.}
}
\value{
Data frame containing station names, station ID codes and dates of
  operation
}
\description{
This function can be used to download a Station Inventory CSV file from
Environment Canada. Note that the 'stations' data set included in this
package contains station data downloaded when the package was last compiled,
so it may not be necessary to call this function (and this function may take a few
minutes to run).
}
