% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations_meta}
\alias{stations_meta}
\title{Show stations list meta data}
\usage{
stations_meta()
}
\description{
Date of ECCC update and date downloaded via weathercan.
}
\examples{
\dontshow{if (check_eccc()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
stations_meta()
\dontshow{\}) # examplesIf}
}
