% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations_search}
\alias{stations_search}
\title{Search for stations by name}
\usage{
stations_search(name = NULL, interval = c("hour", "day", "month"),
  stn = weathercan::stations, verbose = FALSE, quiet = FALSE)
}
\arguments{
\item{name}{Character. A vector of length 1 or more with text against which
to match. Will match station names that contain all components of
\code{name}, but they can be in different orders and separated by other
text.}

\item{interval}{Character. Return only stations with data at these intervals.
Must be any of "hour", "day", "month".}

\item{stn}{Data frame. The \code{stations} data frame to use. Will use the
one included in the package unless otherwise specified.}

\item{verbose}{Logical. Include progress messages}

\item{quiet}{Logical. Suppress all messages (including messages regarding
missing data, etc.)}
}
\value{
Returns a subset of the stations data frame which match the search
  parameters.
}
\description{
Returns stations that match the name provided. This is designed to provide the user with
information with which to decide which station to then get weather data from.
}
\examples{

stations_search(name = "Kamloops")
stations_search(name = "Kamloops", interval = "hour")



}
