% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations_update}
\alias{stations_update}
\title{Update stations data}
\usage{
stations_update()
}
\description{
Runs stations_all() and saves data to the internal \code{weathercan} data
frame: stations. You only need to use this if you need access to new station
information.
}
\details{
The package \code{devtools} is required for this.
}
\examples{

\dontrun{
stations_update()
}

}
