% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables_normals_new}
\alias{variables_normals_new}
\title{Index of variables for new Climate Normals}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{measurement_type}{Measurement category}
\item{ECCC}{Original variable name from ECCC}
\item{weathercan}{R-compatible name given when formatting the data with the
\code{normals_dl()} function}
}
}
\usage{
variables_normals_new
}
\description{
An index matching variables named in weathercan and downloaded with the
\code{normals_dl()} function to those in the original \emph{new} (1991-2020) climate
normals data from ECCC.
}
\keyword{datasets}
