% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables_normals_old}
\alias{variables_normals_old}
\title{Index of variables for new Climate Normals}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{ECCC}{Original variable name from ECCC}
\item{weathercan}{R-compatible name given when formatting the data with the
\code{normals_dl()} function}
}
}
\usage{
variables_normals_old
}
\description{
An index matching variables named in weathercan and downloaded with the
\code{normals_dl()} function to those in the original \emph{old} (1981-2010 and
1971-2000) climate normals data from ECCC.
}
\keyword{datasets}
