% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{weathercan-package}
\alias{weathercan-package}
\alias{weathercan}
\title{Easy downloading of weather data from Environment and Climate Change Canada}
\description{
Extended documentation coming soon... see index for details.
}
\references{
Environment and Climate Change Canada \url{https://www.ec.gc.ca/?lang=En}.
}
