% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_eccc}
\alias{check_eccc}
\title{Check access to ECCC}
\usage{
check_eccc()
}
\value{
FALSE if not, TRUE if so
}
\description{
Checks if \url{https://climate.weather.gc.ca} is available and accessible.
}
\examples{

check_eccc()

}
