% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finches}
\alias{finches}
\title{RFID Data on finch visits to feeders}
\format{An example dataset of finch RFID data for interpolation:
\describe{
  \item{bird_id}{Bird ID number}
  \item{time}{Time}
  \item{feeder_id}{feeder ID}
  \item{species}{Species}
  \item{lat}{Latitude of station location in degree decimal format}
  \item{lon}{Longitude of station location in degree decimal format}
}}
\usage{
finches
}
\description{
RFID Data on finch visits to feeders
}
\keyword{datasets}
