% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kamloops}
\alias{kamloops}
\title{Hourly weather data for Kamloops downloaded with \code{weather()}}
\format{An object of class \code{data.frame} with 4368 rows and 35 columns.}
\usage{
kamloops
}
\description{
An example dataset of hourly weather data for Kamloops
}
\keyword{datasets}
