% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normals.R
\name{normals_dl}
\alias{normals_dl}
\title{Download climate normals from Environment and Climate Change Canada}
\usage{
normals_dl(climate_ids, normals_years = "1981-2010", format = TRUE,
  stn = weathercan::stations, verbose = FALSE, quiet = FALSE)
}
\arguments{
\item{climate_ids}{Character. A vector containing the Climate ID(s) of the
station(s) you wish to download data from. See the \code{\link{stations}}
data frame or the \code{\link{stations_search}} function to find Climate
IDs.}

\item{normals_years}{Character. The year range for which you want climate
normals. Default "1981-2010".}

\item{format}{Logical. If TRUE, formats data for immediate use. If FALSE,
returns data exactly as downloaded from Environment and Climate Change
Canada. Useful for dealing with changes by Environment Canada to the format
of data downloads.}

\item{stn}{Data frame. The \code{stations} data frame to use. Will use the
one included in the package unless otherwise specified.}

\item{verbose}{Logical. Include progress messages}

\item{quiet}{Logical. Suppress all messages (including messages regarding
missing data, etc.)}
}
\value{
tibble with nested normals and first/last frost data
}
\description{
Downloads climate normals from Environment and Climate Change Canada (ECCC)
for one or more stations (defined by \code{climate_id}s). For details and units,
see the glossary-normals vignette (\code{vignette("glossary", package =
"weathercan")}) or the glossary online
\url{http://climate.weather.gc.ca/glossary_e.html}.
}
\details{
Climate normals from ECCC include two types of data, averages by
month for a variety of measurements as well as data relating to the
frost-free period. Because these two data sources are quite different, we
return them as nested data so the user can extract them as they wish.

Climate normals are downloaded from the url stored in
\code{getOption("weathercan.urls.normals")}. To change this location use
\code{options(weathercan.urls.normals = "your_new_url")}.
}
\examples{

# Find the climate_id
stations_search("Brandon A", normals_only = TRUE)

# Download climate normals
n <- normals_dl(climate_ids = "5010480")

# Pull out last frost data
library(tidyr)
f <- unnest(n, "frost")

# Pull out normals
nm <- unnest(n, "data")

# Pull out both (note this can be an awkward data set)
nf <- unnest(n, "frost") \%>\%
  unnest("data")

# Download multiple stations
n <- normals_dl(climate_ids = c("3010234", "3010410", "3010815"))
n

# Note that some have files online but no data
n$data[2]

# Some have no last frost data
n$frost[3]

# To pull out all the data, use 'keep_empty' parameter (Otherwise stations
# missing one or the other will be omitted)
# Note: This requires `tidyr` v1

if(packageVersion("tidyr") >= "1.0.0") {
  unnest(n, "data", keep_empty = TRUE) \%>\%
    unnest("frost", keep_empty = TRUE)
}

# Otherwise, if you don't have/don't want tidyr v1, keep the data separate
nm <- unnest(n, "data")
f <- unnest(n, "frost")

}
