% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normals.R
\name{normals_dl}
\alias{normals_dl}
\title{Download climate normals from Environment and Climate Change Canada}
\usage{
normals_dl(climate_ids, normals_years = "1981-2010", format = TRUE,
  stn = weathercan::stations, url = NULL, verbose = FALSE,
  quiet = FALSE)
}
\arguments{
\item{climate_ids}{Character. A vector containing the Climate ID(s) of the
station(s) you wish to download data from. See the \code{\link{stations}}
data frame or the \code{\link{stations_search}} function to find Climate
IDs.}

\item{normals_years}{Character. The year range for which you want climate
normals. Default "1981-2010".}

\item{format}{Logical. If TRUE, formats data for immediate use. If FALSE,
returns data exactly as downloaded from Environment and Climate Change
Canada. Useful for dealing with changes by Environment Canada to the format
of data downloads.}

\item{stn}{Data frame. The \code{stations} data frame to use. Will use the
one included in the package unless otherwise specified.}

\item{url}{Character. Url from which to grab the data. If NULL uses default
url (see details)}

\item{verbose}{Logical. Include progress messages}

\item{quiet}{Logical. Suppress all messages (including messages regarding
missing data, etc.)}
}
\value{
tibble with nested normals and first/last frost data
}
\description{
Downloads climate normals from Environment and Climate Change Canada (ECCC) for one or
more stations (defined by `climate_id`s). For details and units, see the glossary-normals vignette
(\code{vignette("glossary", package = "weathercan")}) or the glossary online
\url{http://climate.weather.gc.ca/glossary_e.html}.
}
\examples{

normals_dl(climate_ids = c("3010234", "3010410", "3010815"))

normals_dl(climate_ids = "5010480")

}
