% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stations.R
\name{stations_all}
\alias{stations_all}
\title{Get available stations}
\usage{
stations_all(province = "all",
  url = "http://climate.weather.gc.ca/historical_data/search_historic_data_stations_e.html")
}
\arguments{
\item{province}{Character. Single or multiple character vector defining
provinces in two letter codes (\"ON\", \"BC\", etc.) or "all".}

\item{url}{Character. Url from which to grab the station information}
}
\value{
Data frame containing station names, station ID codes and dates of operation
}
\description{
This function can be used to download up-to-date station information from
Environment Canada. Note that the 'stations' data set contains station data
downloaded on June 2nd 2016, so it may not be necessary to call this
function.
}
\examples{


}

