% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tz_calc}
\alias{tz_calc}
\title{DEFUNCT: Get timezone from lat/lon}
\usage{
tz_calc()
}
\description{
Accessed Google API to determine local timezone from coordinates. Defunct
as the API is no longer accessible without a Key.
}
