% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{weather}
\alias{weather}
\title{Download Environment Canada Weather data}
\usage{
weather(station_ids, start = NULL, end = NULL, timeframe = "hour",
  trim = TRUE, avg = "none", best = FALSE, format = TRUE,
  string_as = NA, tz_disp = NULL, stn = NULL,
  url = "http://climate.weather.gc.ca/climate_data/bulk_data_e.html",
  verbose = FALSE)
}
\arguments{
\item{station_ids}{Numeric/Character. A vector containing the ID(s) of the
station(s) you wish to download data from. See the \code{\link{stations}}
data frame or the \code{\link{stations_search}} function to find IDs.}

\item{start}{Date/Character. The start date of the data in YYYY-MM-DD format
(applies to all stations_ids). Defaults to start of range.}

\item{end}{Date/Character. The end date of the data in YYYY-MM-DD format
(applies to all station_ids). Defaults to end of range.}

\item{timeframe}{Character. Timeframe of the data, one of "hour", "day",
"month".}

\item{trim}{Logical. Trim missing values from the start and end of the weather
dataframe.}

\item{avg}{Character. (NOT USEABLE) Whether and how to average the data}

\item{best}{Logical. (NOT USEABLE) If TRUE, returns data at the best frame to
maximize data according to the date range (this could result in less data
from other weather measurements, see Details). If FALSE, returns data from
exact timeframe specified.}

\item{format}{Logical. If TRUE, formats data for immediate use. If FALSE,
returns data exactly as downloaded from Environment Canda. Useful for
dealing with changes by Environment Canada to the format of data downloads.}

\item{string_as}{Character. What value to replace character strings in a
numeric measurement with. See Details.}

\item{tz_disp}{Character. What timezone to display times in (must be one of
\code{OlsonNames()}).}

\item{stn}{Data frame. The \code{stations} data frame to use. Will use the one
included in the package unless otherwise specified.}

\item{url}{Character. Url from which to grab the weather data}

\item{verbose}{Logical. Include messages}
}
\value{
Data frame with station ID, name and weather data.
}
\description{
Downloads data from Environment Canada for one or more stations.
}
\details{
Data can be returned 'raw' (format = FALSE) or can be formatted.
Formatting transforms dates/times to date/time class, renames columns, and
converts data to numeric where possible. If character strings are contained in
traditionally numeric fields (e.g., weather speed may have values such as "<
30"), they can be replaced with a character specified by `string_as`. The
default is NA. Formating also replaces data associated with certain flags
with NA (M = Missing).

Start and end date can be specified, but if not, it will
default to the start and end date of the range (this could result in
downloading a lot of data!).

Times are returned as the Etc/GMT offset timezone corresponding to the
location. This doesn't include daylight savings. However, for compatibility
with other data sets, timezones can be converted by specifying the desired
timezone in `tz_disp`.
}
\examples{

\dontrun{
kam <- weather(station_ids = 51423,
               start = "2016-01-01", end = "2016-02-15")
}

stations_search("Kamloops A$", timeframe = "hour")
stations_search("Prince George Airport", timeframe = "hour")
\dontrun{
kam.pg <- weather(station_ids = c(48248, 51423),
                  start = "2016-01-01", end = "2016-02-15")

library(ggplot2)

ggplot(data = kam.pg, aes(x = time, y = temp, group = station_name, colour = station_name)) +
       geom_line()
}


}

