% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{weather}
\alias{weather}
\title{Download Environment Canada Weather data}
\usage{
weather(station_ID, start = NULL, end = NULL, by = "hour", avg = "none",
  best = TRUE, format = TRUE, string_as_na = TRUE,
  url = "http://climate.weather.gc.ca/climate_data/bulk_data_e.html")
}
\arguments{
\item{url}{}
}
\description{
First check for interval and timeframe specified. Then if that doesn't match,
prioritorize according to the following: - Maximize the amount of time
covered by the interval: Download by smaller time frames to allow averaging
(override with '\code{best = FALSE}') - Move the start/end dates to the
min/max ranges available
}

